+++ tests/Equinox.EventStore.Integration/StoreIntegration.fs
@@ -8,12 +8,17 @@ open Swensen.Unquote
 open System.Threading
 open System
 
+let defaultBatchSize = 500
+
 #if STORE_POSTGRES
 open Equinox.SqlStreamStore
 open Equinox.SqlStreamStore.Postgres
 
 let connectToLocalStore (_ : ILogger) =
     Connector("Host=localhost;User Id=postgres;database=EQUINOX_TEST_DB",autoCreate=true).Establish()
+
+type Context = SqlStreamStoreContext
+type Category<'event, 'state, 'context> = SqlStreamStoreCategory<'event, 'state, 'context>
 #else
 #if STORE_MSSQL
 open Equinox.SqlStreamStore
@@ -21,6 +26,9 @@ open Equinox.SqlStreamStore.MsSql
 
 let connectToLocalStore (_ : ILogger) =
     Connector(sprintf "Server=localhost,1433;User=sa;Password=!Passw0rd;Database=test",autoCreate=true).Establish()
+
+type Context = SqlStreamStoreContext
+type Category<'event, 'state, 'context> = SqlStreamStoreCategory<'event, 'state, 'context>
 #else
 #if STORE_MYSQL
 open Equinox.SqlStreamStore
@@ -28,6 +36,9 @@ open Equinox.SqlStreamStore.MySql
 
 let connectToLocalStore (_ : ILogger) =
     Connector(sprintf "Server=localhost;User=root;Database=EQUINOX_TEST_DB",autoCreate=true).Establish()
+
+type Context = SqlStreamStoreContext
+type Category<'event, 'state, 'context> = SqlStreamStoreCategory<'event, 'state, 'context>
 #else // STORE_EVENTSTORE
 open Equinox.EventStore
 
@@ -39,39 +50,41 @@ open Equinox.EventStore
 let connectToLocalStore log =
     Connector("admin", "changeit", reqTimeout=TimeSpan.FromSeconds 3., reqRetries=3, log=Logger.SerilogVerbose log, tags=["I",Guid.NewGuid() |> string])
         .Establish("Equinox-integration", Discovery.Uri(Uri "tcp://localhost:1113"),ConnectionStrategy.ClusterSingle NodePreference.Master)
+
+type Context = EventStoreContext
+type Category<'event, 'state, 'context> = EventStoreCategory<'event, 'state, 'context>
 #endif
 #endif
 #endif
 
-let defaultBatchSize = 500
-let createGesGateway connection batchSize = Context(connection, BatchingPolicy(maxBatchSize = batchSize))
+let createContext connection batchSize = Context(connection, BatchingPolicy(maxBatchSize = batchSize))
 
 module Cart =
     let fold, initial = Domain.Cart.Fold.fold, Domain.Cart.Fold.initial
     let codec = Domain.Cart.Events.codec
     let snapshot = Domain.Cart.Fold.isOrigin, Domain.Cart.Fold.snapshot
-    let createServiceWithoutOptimization log gateway =
-        let resolve (id,opt) = Resolver(gateway, Domain.Cart.Events.codec, fold, initial).Resolve(id,?option=opt)
+    let createServiceWithoutOptimization log context =
+        let resolve (id,opt) = Category(context, Domain.Cart.Events.codec, fold, initial).Resolve(id,?option=opt)
         Cart.create log resolve
-    let createServiceWithCompaction log gateway =
-        let resolve (id,opt) = Resolver(gateway, codec, fold, initial, access = AccessStrategy.RollingSnapshots snapshot).Resolve(id,?option=opt)
+    let createServiceWithCompaction log context =
+        let resolve (id,opt) = Category(context, codec, fold, initial, access = AccessStrategy.RollingSnapshots snapshot).Resolve(id,?option=opt)
         Cart.create log resolve
-    let createServiceWithCaching log gateway cache =
+    let createServiceWithCaching log context cache =
         let sliding20m = CachingStrategy.SlidingWindow (cache, TimeSpan.FromMinutes 20.)
-        Cart.create log (fun (id,opt) -> Resolver(gateway, codec, fold, initial, sliding20m).Resolve(id,?option=opt))
-    let createServiceWithCompactionAndCaching log gateway cache =
+        Cart.create log (fun (id,opt) -> Category(context, codec, fold, initial, sliding20m).Resolve(id,?option=opt))
+    let createServiceWithCompactionAndCaching log context cache =
         let sliding20m = CachingStrategy.SlidingWindow (cache, TimeSpan.FromMinutes 20.)
-        Cart.create log (fun (id,opt) -> Resolver(gateway, codec, fold, initial, sliding20m, AccessStrategy.RollingSnapshots snapshot).Resolve(id,?option=opt))
+        Cart.create log (fun (id,opt) -> Category(context, codec, fold, initial, sliding20m, AccessStrategy.RollingSnapshots snapshot).Resolve(id,?option=opt))
 
 module ContactPreferences =
     let fold, initial = Domain.ContactPreferences.Fold.fold, Domain.ContactPreferences.Fold.initial
     let codec = Domain.ContactPreferences.Events.codec
     let createServiceWithoutOptimization log connection =
-        let gateway = createGesGateway connection defaultBatchSize
-        ContactPreferences.create log (Resolver(gateway, codec, fold, initial).Resolve)
+        let context = createContext connection defaultBatchSize
+        ContactPreferences.create log (Category(context, codec, fold, initial).Resolve)
     let createService log connection =
-        let resolver = Resolver(createGesGateway connection 1, codec, fold, initial, access = AccessStrategy.LatestKnownEvent)
-        ContactPreferences.create log resolver.Resolve
+        let cat = Category(createContext connection 1, codec, fold, initial, access = AccessStrategy.LatestKnownEvent)
+        ContactPreferences.create log cat.Resolve
 
 #nowarn "1182" // From hereon in, we may have some 'unused' privates (the tests)
 
@@ -106,19 +119,19 @@ type Tests(testOutputHelper) =
     let batchForwardAndAppend = singleBatchForward @ [EsAct.Append]
 
     [<AutoData(SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_EVENTSTORE")>]
-    let ``Can roundtrip against EventStore, correctly batching the reads [without any optimizations]`` context skuId = Async.RunSynchronously <| async {
+    let ``Can roundtrip against EventStore, correctly batching the reads [without any optimizations]`` (ctx, skuId) = Async.RunSynchronously <| async {
         let log, capture = createLoggerWithCapture ()
-        let! conn = connectToLocalStore log
+        let! connection = connectToLocalStore log
 
         let batchSize = 3
-        let gateway = createGesGateway conn batchSize
-        let service = Cart.createServiceWithoutOptimization log gateway
+        let context = createContext connection batchSize
+        let service = Cart.createServiceWithoutOptimization log context
 
         // The command processing should trigger only a single read and a single write call
         let addRemoveCount = 6
         let cartId = % System.Guid.NewGuid()
 
-        do! addAndThenRemoveItemsManyTimesExceptTheLastOne context cartId skuId service addRemoveCount
+        do! addAndThenRemoveItemsManyTimesExceptTheLastOne ctx cartId skuId service addRemoveCount
         test <@ batchForwardAndAppend = capture.ExternalCalls @>
 
         // Restart the counting
@@ -137,28 +150,27 @@ type Tests(testOutputHelper) =
     [<AutoData(MaxTest = 2, SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_EVENTSTORE")>]
     let ``Can roundtrip against EventStore, managing sync conflicts by retrying [without any optimizations]`` (ctx, initialState) = Async.RunSynchronously <| async {
         let log1, capture1 = createLoggerWithCapture ()
-        let! conn = connectToLocalStore log1
+        let! connection = connectToLocalStore log1
         // Ensure batching is included at some point in the proceedings
         let batchSize = 3
 
-        let context, (sku11, sku12, sku21, sku22) = ctx
+        let ctx, (sku11, sku12, sku21, sku22) = ctx
         let cartId = % System.Guid.NewGuid()
 
-
         // establish base stream state
-        let gateway = createGesGateway conn batchSize
-        let service1 = Cart.createServiceWithoutOptimization log1 gateway
+        let context = createContext connection batchSize
+        let service1 = Cart.createServiceWithoutOptimization log1 context
         let! maybeInitialSku =
             let (streamEmpty, skuId) = initialState
             async {
                 if streamEmpty then return None
                 else
                     let addRemoveCount = 2
-                    do! addAndThenRemoveItemsManyTimesExceptTheLastOne context cartId skuId service1 addRemoveCount
+                    do! addAndThenRemoveItemsManyTimesExceptTheLastOne ctx cartId skuId service1 addRemoveCount
                     return Some (skuId, addRemoveCount) }
 
         let act prepare (service : Cart.Service) log skuId count =
-            service.ExecuteManyAsync(cartId, false, prepare = prepare, commands = [Domain.Cart.SyncItem (context, skuId, Some count, None)])
+            service.ExecuteManyAsync(cartId, false, prepare = prepare, commands = [Domain.Cart.SyncItem (ctx, skuId, Some count, None)])
 
         let eventWaitSet () = let e = new ManualResetEvent(false) in (Async.AwaitWaitHandle e |> Async.Ignore), async { e.Set() |> ignore }
         let w0, s0 = eventWaitSet ()
@@ -179,8 +191,8 @@ type Tests(testOutputHelper) =
             // Signal conflict generated
             do! s4 }
         let log2, capture2 = createLoggerWithCapture ()
-        let gateway = createGesGateway conn batchSize
-        let service2 = Cart.createServiceWithoutOptimization log2 gateway
+        let context = createContext connection batchSize
+        let service2 = Cart.createServiceWithoutOptimization log2 context
         let t2 = async {
             // Signal we have state, wait for other to do same, engineer conflict
             let prepare = async {
@@ -214,16 +226,16 @@ type Tests(testOutputHelper) =
     let batchBackwardsAndAppend = singleBatchBackwards @ [EsAct.Append]
 
     [<AutoData(SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_EVENTSTORE")>]
-    let ``Can roundtrip against EventStore, correctly compacting to avoid redundant reads`` context skuId = Async.RunSynchronously <| async {
+    let ``Can roundtrip against EventStore, correctly compacting to avoid redundant reads`` (ctx, skuId) = Async.RunSynchronously <| async {
         let log, capture = createLoggerWithCapture ()
-        let! conn = connectToLocalStore log
+        let! client = connectToLocalStore log
         let batchSize = 10
-        let gateway = createGesGateway conn batchSize
-        let service = Cart.createServiceWithCompaction log gateway
+        let context = createContext client batchSize
+        let service = Cart.createServiceWithCompaction log context
 
         // Trigger 10 events, then reload
         let cartId = % System.Guid.NewGuid()
-        do! addAndThenRemoveItemsManyTimes context cartId skuId service 5
+        do! addAndThenRemoveItemsManyTimes ctx cartId skuId service 5
         let! _ = service.Read cartId
 
         // ... should see a single read as we are inside the batch threshold
@@ -231,7 +243,7 @@ type Tests(testOutputHelper) =
 
         // Add two more, which should push it over the threshold and hence trigger inclusion of a snapshot event (but not incurr extra roundtrips)
         capture.Clear()
-        do! addAndThenRemoveItemsManyTimes context cartId skuId service 1
+        do! addAndThenRemoveItemsManyTimes ctx cartId skuId service 1
         test <@ batchBackwardsAndAppend = capture.ExternalCalls @>
 
         // While we now have 13 events, we should be able to read them with a single call
@@ -241,14 +253,14 @@ type Tests(testOutputHelper) =
 
         // Add 8 more; total of 21 should not trigger snapshotting as Event Number 12 (the 13th one) is a shapshot
         capture.Clear()
-        do! addAndThenRemoveItemsManyTimes context cartId skuId service 4
+        do! addAndThenRemoveItemsManyTimes ctx cartId skuId service 4
         test <@ batchBackwardsAndAppend = capture.ExternalCalls @>
 
         // While we now have 21 events, we should be able to read them with a single call
         capture.Clear()
         let! _ = service.Read cartId
         // ... and trigger a second snapshotting (inducing a single additional read + write)
-        do! addAndThenRemoveItemsManyTimes context cartId skuId service 1
+        do! addAndThenRemoveItemsManyTimes ctx cartId skuId service 1
         // and reload the 24 events with a single read
         let! _ = service.Read cartId
         test <@ singleBatchBackwards @ batchBackwardsAndAppend @ singleBatchBackwards = capture.ExternalCalls @>
@@ -257,8 +269,8 @@ type Tests(testOutputHelper) =
     [<AutoData(SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_EVENTSTORE")>]
     let ``Can correctly read and update against EventStore, with LatestKnownEvent Access Strategy`` id value = Async.RunSynchronously <| async {
         let log, capture = createLoggerWithCapture ()
-        let! conn = connectToLocalStore log
-        let service = ContactPreferences.createService log conn
+        let! client = connectToLocalStore log
+        let service = ContactPreferences.createService log client
 
         // Feed some junk into the stream
         for i in 0..11 do
@@ -277,18 +289,18 @@ type Tests(testOutputHelper) =
     }
 
     [<AutoData(SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_EVENTSTORE")>]
-    let ``Can roundtrip against EventStore, correctly caching to avoid redundant reads`` context skuId = Async.RunSynchronously <| async {
+    let ``Can roundtrip against EventStore, correctly caching to avoid redundant reads`` (ctx, skuId) = Async.RunSynchronously <| async {
         let log, capture = createLoggerWithCapture ()
-        let! conn = connectToLocalStore log
+        let! client = connectToLocalStore log
         let batchSize = 10
         let cache = Equinox.Cache("cart", sizeMb = 50)
-        let gateway = createGesGateway conn batchSize
-        let createServiceCached () = Cart.createServiceWithCaching log gateway cache
-        let service1, service2, service3 = createServiceCached (), createServiceCached (), Cart.createServiceWithoutOptimization log gateway
+        let context = createContext client batchSize
+        let createServiceCached () = Cart.createServiceWithCaching log context cache
+        let service1, service2, service3 = createServiceCached (), createServiceCached (), Cart.createServiceWithoutOptimization log context
         let cartId = % Guid.NewGuid()
 
         // Trigger 10 events, then reload
-        do! addAndThenRemoveItemsManyTimesExceptTheLastOne context cartId skuId service1 5
+        do! addAndThenRemoveItemsManyTimesExceptTheLastOne ctx cartId skuId service1 5
         test <@ batchForwardAndAppend = capture.ExternalCalls @>
         let! resStale = service2.ReadStale cartId
         test <@ batchForwardAndAppend = capture.ExternalCalls @>
@@ -301,7 +313,7 @@ type Tests(testOutputHelper) =
         // Add two more - the roundtrip should only incur a single read
         capture.Clear()
         let skuId2 = SkuId <| Guid.NewGuid()
-        do! addAndThenRemoveItemsManyTimesExceptTheLastOne context cartId skuId2 service1 1
+        do! addAndThenRemoveItemsManyTimesExceptTheLastOne ctx cartId skuId2 service1 1
         test <@ batchForwardAndAppend = capture.ExternalCalls @>
 
         // While we now have 12 events, we should be able to read them with a single call
@@ -318,41 +330,41 @@ type Tests(testOutputHelper) =
         // Optimistic transactions
         capture.Clear()
         // As the cache is up to date, we can transact against the cached value and do a null transaction without a roundtrip
-        do! addAndThenRemoveItemsOptimisticManyTimesExceptTheLastOne context cartId skuId2 service1 1
+        do! addAndThenRemoveItemsOptimisticManyTimesExceptTheLastOne ctx cartId skuId2 service1 1
         test <@ [] = capture.ExternalCalls @>
         // As the cache is up to date, we can do an optimistic append, saving a Read roundtrip
         let skuId3 = SkuId <| Guid.NewGuid()
-        do! addAndThenRemoveItemsOptimisticManyTimesExceptTheLastOne context cartId skuId3 service1 1
+        do! addAndThenRemoveItemsOptimisticManyTimesExceptTheLastOne ctx cartId skuId3 service1 1
         // this time, we did something, so we see the append call
         test <@ [EsAct.Append] = capture.ExternalCalls @>
 
         // If we don't have a cache attached, we don't benefit from / pay the price for any optimism
         capture.Clear()
         let skuId4 = SkuId <| Guid.NewGuid()
-        do! addAndThenRemoveItemsOptimisticManyTimesExceptTheLastOne context cartId skuId4 service3 1
+        do! addAndThenRemoveItemsOptimisticManyTimesExceptTheLastOne ctx cartId skuId4 service3 1
         // Need 2 batches to do the reading
         test <@ [EsAct.SliceForward] @ singleBatchForward @ [EsAct.Append] = capture.ExternalCalls @>
         // we've engineered a clash with the cache state (service3 doest participate in caching)
         // Conflict with cached state leads to a read forward to resync; Then we'll idempotently decide not to do any append
         capture.Clear()
-        do! addAndThenRemoveItemsOptimisticManyTimesExceptTheLastOne context cartId skuId4 service2 1
+        do! addAndThenRemoveItemsOptimisticManyTimesExceptTheLastOne ctx cartId skuId4 service2 1
         test <@ [EsAct.AppendConflict; EsAct.SliceForward; EsAct.BatchForward] = capture.ExternalCalls @>
     }
 
     [<AutoData(SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_EVENTSTORE")>]
-    let ``Can combine compaction with caching against EventStore`` context skuId = Async.RunSynchronously <| async {
+    let ``Can combine compaction with caching against EventStore`` (ctx, skuId) = Async.RunSynchronously <| async {
         let log, capture = createLoggerWithCapture ()
-        let! conn = connectToLocalStore log
+        let! client = connectToLocalStore log
         let batchSize = 10
-        let gateway = createGesGateway conn batchSize
-        let service1 = Cart.createServiceWithCompaction log gateway
+        let context = createContext client batchSize
+        let service1 = Cart.createServiceWithCompaction log context
         let cache = Equinox.Cache("cart", sizeMb = 50)
-        let gateway = createGesGateway conn batchSize
-        let service2 = Cart.createServiceWithCompactionAndCaching log gateway cache
+        let context = createContext client batchSize
+        let service2 = Cart.createServiceWithCompactionAndCaching log context cache
 
         // Trigger 10 events, then reload
         let cartId = % System.Guid.NewGuid()
-        do! addAndThenRemoveItemsManyTimes context cartId skuId service1 5
+        do! addAndThenRemoveItemsManyTimes ctx cartId skuId service1 5
         let! _ = service2.Read cartId
 
         // ... should see a single read as we are inside the batch threshold
@@ -360,7 +372,7 @@ type Tests(testOutputHelper) =
 
         // Add two more, which should push it over the threshold and hence trigger inclusion of a snapshot event (but not incur extra roundtrips)
         capture.Clear()
-        do! addAndThenRemoveItemsManyTimes context cartId skuId service1 1
+        do! addAndThenRemoveItemsManyTimes ctx cartId skuId service1 1
         test <@ batchBackwardsAndAppend = capture.ExternalCalls @>
 
         // While we now have 13 events, we should be able to read them backwards with a single call
@@ -370,14 +382,14 @@ type Tests(testOutputHelper) =
 
         // Add 8 more; total of 21 should not trigger snapshotting as Event Number 12 (the 13th one) is a snapshot
         capture.Clear()
-        do! addAndThenRemoveItemsManyTimes context cartId skuId service1 4
+        do! addAndThenRemoveItemsManyTimes ctx cartId skuId service1 4
         test <@ batchBackwardsAndAppend = capture.ExternalCalls @>
 
         // While we now have 21 events, we should be able to read them with a single call
         capture.Clear()
         let! _ = service1.Read cartId
         // ... and trigger a second snapshotting (inducing a single additional read + write)
-        do! addAndThenRemoveItemsManyTimes context cartId skuId service1 1
+        do! addAndThenRemoveItemsManyTimes ctx cartId skuId service1 1
         // and we _could_ reload the 24 events with a single read if reading backwards. However we are using the cache, which last saw it with 10 events, which necessitates two reads
         let! _ = service2.Read cartId
         let suboptimalExtraSlice = [singleSliceForward]