+++ tests/Equinox.MemoryStore.Integration/MemoryStoreIntegration.fs
@@ -6,8 +6,8 @@ open Swensen.Unquote
 
 let createMemoryStore () = VolatileStore<_>()
 let createServiceMemory log store =
-    let resolver = Resolver(store, FsCodec.Box.Codec.Create(), Domain.Cart.Fold.fold, Domain.Cart.Fold.initial)
-    let resolve (id, opt) = resolver.Resolve(id, ?option=opt)
+    let cat = MemoryStoreCategory(store, FsCodec.Box.Codec.Create(), Domain.Cart.Fold.fold, Domain.Cart.Fold.initial)
+    let resolve (id, opt) = cat.Resolve(id, ?option=opt)
     Cart.create log resolve
 
 #nowarn "1182" // From hereon in, we may have some 'unused' privates (the tests)
@@ -51,8 +51,8 @@ type Tests(testOutputHelper) =
     }
 
 let createFavoritesServiceMemory log store =
-    let resolver = Resolver(store, FsCodec.Box.Codec.Create(), Domain.Favorites.Fold.fold, Domain.Favorites.Fold.initial)
-    Favorites.create log resolver.Resolve
+    let cat = MemoryStoreCategory(store, FsCodec.Box.Codec.Create(), Domain.Favorites.Fold.fold, Domain.Favorites.Fold.initial)
+    Favorites.create log cat.Resolve
 
 type ChangeFeed(testOutputHelper) =
     let testOutput = TestOutputAdapter testOutputHelper